Rename max_reduce_prod_join feature header and source file
