+++ searchlib/src/tests/features/internal_max_reduce_prod_join_feature/internal_max_reduce_prod_join_feature_test.cpp
@@ -4,7 +4,7 @@
 
 #include <vespa/searchlib/attribute/attributefactory.h>
 #include <vespa/searchlib/fef/test/ftlib.h>
-#include <vespa/searchlib/features/max_reduce_prod_join_feature.h>
+#include <vespa/searchlib/features/internal_max_reduce_prod_join_feature.h>
 #include <vespa/searchlib/attribute/attribute.h>
 
 using search::feature_t;
