+++ searchlib/src/vespa/searchlib/features/internal_max_reduce_prod_join_feature.cpp
@@ -1,6 +1,6 @@
 // Copyright 2017 Yahoo Holdings. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 
-#include "max_reduce_prod_join_feature.h"
+#include "internal_max_reduce_prod_join_feature.h"
 #include "valuefeature.h"
 #include "weighted_set_parser.h"
