+++ searchlib/src/vespa/searchlib/features/internal_max_reduce_prod_join_feature.h
